pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyclaw"
version = "2.0.0"
description = "OpenClaw 的 Python 实现 - 基于 LangGraph 的多 Agent 协作系统"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "PyClaw Team", email = "pyclaw@example.com"}
]
maintainers = [
    {name = "PyClaw Team", email = "pyclaw@example.com"}
]
keywords = [
    "llm",
    "agent",
    "langchain",
    "langgraph",
    "rag",
    "ai",
    "chatbot",
    "automation",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"

# 核心依赖
dependencies = [
    # WebSocket 和 HTTP
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "requests>=2.32.0",
    
    # CLI 和配置
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "json5>=0.9.14",
    "pyyaml>=6.0",
    
    # LangChain 核心 (0.3+ 架构)
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    
    # LangChain 社区和 Provider
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    
    # LLM Provider SDK
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    
    # 向量记忆
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
    
    # 网页工具
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
# 性能优化
performance = [
    "faiss-cpu>=1.7.4",      # 向量检索加速
    "orjson>=3.9.0",         # 更快的 JSON 解析
]

# 渠道集成
channels = [
    "python-telegram-bot>=20.0",
    "discord.py>=2.3.0",
]

# 开发工具
dev = [
    # 测试
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    
    # 代码质量
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    
    # 性能分析
    "memory-profiler>=0.60.0",
    "line-profiler>=4.0.0",
    
    # 文档
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=2.0.0",
    
    # 监控
    "prometheus-client>=0.17.0",
]

# 完整安装（所有依赖）
all = [
    "pyclaw[performance,channels,dev]",
]

[project.urls]
Homepage = "https://github.com/pyclaw/pyclaw"
Documentation = "https://pyclaw.readthedocs.io"
Repository = "https://github.com/pyclaw/pyclaw.git"
Issues = "https://github.com/pyclaw/pyclaw/issues"
Changelog = "https://github.com/pyclaw/pyclaw/blob/main/CHANGELOG.md"

[project.scripts]
pyclaw = "pyclaw.cli:main"

[project.entry-points."pyclaw.plugins"]
# 预留插件入口点
# example = "pyclaw.plugins.example:ExamplePlugin"

[tool.setuptools]
packages = ["pyclaw"]
include-package-data = true

[tool.setuptools.package-data]
pyclaw = ["*.md", "*.json", "*.yaml"]

[tool.setuptools.exclude-package-data]
pyclaw = ["tests/*", "*.pyc", "__pycache__/*"]

# Black 配置
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# isort 配置
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# pytest 配置
[tool.pytest.ini_options]
minversion = "7.4"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=pyclaw",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# mypy 配置
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

# 覆盖特定模块
[[tool.mypy.overrides]]
module = "pyclaw.*"
disallow_untyped_defs = true

# flake8 配置
[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".eggs",
    "*.egg",
]
per-file-ignores = [
    "__init__.py:F401",
]
